% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.r
\name{as.distributional}
\alias{as.distributional}
\title{Convert a ragged list into a \code{distributional} object.}
\usage{
as.distributional(x, err = NULL, method = c("KS", "SH"), xlabel = "Ma",
  range = NULL)
}
\arguments{
\item{x}{List of values.}

\item{err}{Optional list or errors.}

\item{method}{Dissimilarity metric.}

\item{xlabel}{Data label.}

\item{range}{Data range.}
}
\value{
A \code{distributional} object for use with package \code{provenance}.
}
\description{
Convert a ragged list into a \code{distributional} object.
}
\keyword{internal}

