% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendrogram.r
\name{clustree}
\alias{clustree}
\title{Parse a hclust structure for plotting}
\usage{
clustree(atree, orientation = c(1:4), type = c(1, 2), positions = NULL,
  classes = 1)
}
\arguments{
\item{atree}{An object of class \code{hclust}.}

\item{orientation}{Branching direction: 1 top-down, 2 right-to-left, 3
bottom-up, 4 left-to-right.}

\item{type}{1 uniform height steps, 2 calculated by hclust with base = 0, 3
same as 2, with base = height from hclust.}

\item{positions}{Takes 'y'-positions in the order of drawing layout
(atree$order), i.e. ascending values.}

\item{classes}{Cut tree into sub-branches, return as column in $segments
and $leaves.}
}
\value{
List of geometric elements for plotting the tree structure and
labels.
}
\description{
Parse a hclust structure for plotting
}
\details{
Takes a hclust structure and returns rendered coordinates to draw
the tree, according to \code{orientation} and \code{type} parameters. Also
returns leaf labels. Generates segments of tree based on pre-calculated
coordinates for leaves and in any orientation, unlike \code{plot.hclust} or
package \code{ggdendro}.
}
\seealso{
\code{\link[stats]{hclust}}
}
\keyword{internal}

