% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrovenance-package.r
\docType{package}
\name{ggprovenance-package}
\alias{ggprovenance-package}
\title{ggprovenance: Plotting of sediment provenance data, ggplot-style.}
\description{
Plotting functions wrapping extended and user friendly graphical output
around the calculations provided by package provenance. Plots are returned
as ggplot objects, and can be further modified/combined with ggplot2's
functions, which also provide export to many file formats. Helper functions
to easily read common formats of input data are provided (requires ggdat).
}
\section{Main functions}{


\code{\link[ggprovenance]{plotKDE}}: Plotting of kernel density estimates
(KDEs) of distributional data, typically single-grain ages.

\code{\link[ggprovenance]{plotMDS}}: Multi-dimensional scaling (MDS) maps
of sample relationships, of either distributional or categorical data, e.g.
single-grain ages or heavy mineral counts, respectively.
}

