% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendrogram.r
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Plot a dendrogram}
\usage{
plotDendrogram(data, orientation = 4, type = 2,
  positions = c(1:length(data)), metric = c("KS", "SH", "aitchinson",
  "bray"), method = "ward.D", classes = 1)
}
\arguments{
\item{data}{Data to compare.}

\item{orientation}{See \code{\link{clustree}}.}

\item{type}{See \code{\link{clustree}}.}

\item{positions}{See \code{\link{clustree}}.}

\item{metric}{Metric to compare samples within data.}

\item{method}{Agglomeration method, see \code{\link[stats]{hclust}}.}

\item{classes}{Cut tree into branches.}
}
\value{
A ggplot2 object containing the specified dendrogram.
}
\description{
Plot a dendrogram
}
\seealso{
\code{\link{clustree}}, \code{\link[stats]{hclust}}, \code{\link[provenance]{diss}}
}

