% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.r
\name{plotMDS}
\alias{plotMDS}
\title{Plotting multi-dimensional scaling maps.}
\usage{
plotMDS(x, diss, categories, mapping, nearest = TRUE, fixed = FALSE,
  axes = FALSE, square = FALSE, metric = c("K-S", "C-v-M"))
}
\arguments{
\item{x}{Input data. Optional (see \code{diss}.}

\item{diss}{Optional distances/dissimilarity matrix. Mandatory if \code{x}
is missing.}

\item{categories}{A data frame giving categorising information on samples.}

\item{mapping}{Aesthetic mapping, see \code{\link[ggplot2]{aes}}.}

\item{nearest}{Plot lines connecting nearest distances.}

\item{fixed}{Plot in fixed (1:1) aspect ratio.}

\item{axes}{Plot axes tick marks.}

\item{square}{Expand plot area to yield square plot.}

\item{metric}{Metric to use in dissimilarity calculation.}
}
\value{
A ggplot object containing the specified plot.
}
\description{
Plotting multi-dimensional scaling maps.
}

