% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{prettyBreaks}
\alias{prettyBreaks}
\title{Return "pretty" breaks for axis labelling.}
\usage{
prettyBreaks(range, inner = TRUE, logsc = FALSE, majors = TRUE,
  minors = FALSE, n.major = 8, n.minor = 0, force.limits = FALSE,
  frac.log = FALSE)
}
\arguments{
\item{range}{Numeric vector of length 2 giving the min and max input values.}

\item{inner}{\code{TRUE}: return only "pretty" ticks within \code{range};
\code{FALSE}: return ticks spanning closest "pretty" ticks outside
\code{range}.}

\item{logsc}{Boolean, return values suitable for labelling log-scale axes.}

\item{majors}{Boolean, return major ticks.}

\item{minors}{Boolean, return minor ticks.}

\item{n.major}{(Approximate) number of major ticks to return.}

\item{n.minor}{Number of minor ticks to return between each pair of major
ticks.}

\item{force.limits}{If \code{inner = TRUE}, also include exact \code{range}
values.}

\item{frac.log}{Use fractions of log10 (10^0.5, 10^0.33, 10^0.25), to closer
match desired number of ticks.}
}
\value{
A numerical vector of generated breaks.
}
\description{
This function is similar to \code{pretty_breaks} in package
\code{scales}, but has advanced options for log scales, forced range limits,
and separate generation of major and minor ticks (useful for labelling).
}
\details{
Setting \code{major = FALSE, minor=TRUE} can be useful to return
minor ticks only, e.g. to plot them in a different style than major ticks.
}
\keyword{internal}

