% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.r
\name{read.xls.flat}
\alias{read.xls.flat}
\title{Read a 'flat' Excel age data file}
\usage{
read.xls.flat(filename, ...)
}
\arguments{
\item{filename}{The filename to load.}

\item{...}{Additional parameters passed on to \code{read.xls}.}
}
\value{
A list of distributional data vectors (like e.g. ages) of varying
length. The names are taken from the column names in the first line in the
data file, usually sample names.
}
\description{
Reads distributional data (like measured ages) from an Excel
file in which all data is contained in one worksheet, one column per sample.
The sample names are taken from the first line (headers).
}
\details{
Further arguments to \code{read.xls()} can be passed in the
\code{...} parameter, e.g. \code{skip=n} to skip n lines at the top of the
table, or \code{sheet="abc"} to specify the name of the worksheet containing
the data.
}
\examples{
datafile<-system.file("extdata", "Tarim.xls", package="ggprovenance")
agedata<-read.xls.flat(datafile)
}
\seealso{
\code{\link{read.xls.tabbed}}, \code{\link[gdata]{read.xls}}
}
\keyword{internal}

